ghostlink-backend/requirements.txt
axios==0.4.0
backports.tarfile==1.2.0
flask-cors==4.0.1
flask-restful==0.3.10
google-cloud-storage==2.18.2
google-cloud-videointelligence==2.13.5
grpcio-status==1.66.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
jaraco.text==3.12.1
moviepy==1.0.3
opencv-python==4.10.0.84
pip-chill==1.0.3
platformdirs==4.2.2
pytesseract==0.3.13
python-dotenv==1.0.1
speechrecognition==3.10.4
tomli==2.0.1
transformers==4.44.2
